refactor. type 지정 파일 구조 변경
