--- src/types/home.ts
@@ -1,4 +1,0 @@
-export interface HeaderProps {
-  name: string;
-  type: string;
-}
