--- src/types/layout.ts
@@ -1,3 +1,0 @@
-export interface LayoutProps {
-  children: JSX.Element;
-}
