--- src/types/menu.ts
@@ -1,3 +1,0 @@
-export interface MenuLinkProps {
-  name: string;
-}
