+++ src/types/types.ts
@@ -1,0 +1,12 @@
+export interface HeaderProps {
+  name: string;
+  type: string;
+}
+
+export interface LayoutProps {
+  children: JSX.Element;
+}
+
+export interface MenuLinkProps {
+  name: string;
+}
